Release 1.0.1: minor bug fix
Fixed a bug where the repository summary could be confusing if an error
was encountered.

Strictly speaking this is a breaking API change because we now set
`#[must_use]` on many functions. However, we update the version number
based on changes to the command line tool, not the crate API.