Use `fmt::Display` instead of `AsRef<str>`
In places where we convert to a `String` with `to_string()` anyway, we
might as well use a trait that provides that rather than one that allows
us to get a `str`.

For some reason using the `ToString` trait causes clippy to complain
about not using references, whereas using `fmt::Display` (which implies
the `ToString` trait) does not.

This also adds a note in the CHANGELOG.md about updating
`ShellWriter::with_prefix()`.