Release 1.0.3: security updates and breaking API changes
This release updates a few dependencies to avoid security
vulnerabilities that probably did not affect git-status-vars.

It also cleans up the API in a number of small ways. This is a breaking
change to the API, but as stated in DEVELOPMENT.md, the version number
tracks the command line utility, not the API. In the unlikely event that
someone is using the API, the changes are fairly trivial.