Add a basic timeout (1 second by default)
This uses [`alarm()`] to kill the `git-status-vars` process if it takes
too long.

This switches to outputting repo information as soon as possible, but
only outputting `repo_state=` at the end — if an error is encountered,
it can still be reported through `repo_state`.

This adds `unsafe` code to the binary, but not to the library crate.

Fixes: #67 — Needs timeout or some way of dealing with slowness

[`alarm()`]: https://man7.org/linux/man-pages/man2/alarm.2.html